+++ server/incore-common/src/main/java/edu/illinois/ncsa/incore/common/models/GroupAllocations.java
@@ -14,9 +14,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.morphia.annotations.Entity;
 import dev.morphia.annotations.Id;
 import dev.morphia.annotations.Property;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import org.bson.types.ObjectId;
-
-import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 @Entity("GroupAllocations")