+++ server/space-service/src/main/java/edu/illinois/ncsa/incore/service/space/GenericExceptionMapper.java
@@ -10,12 +10,11 @@
 
 package edu.illinois.ncsa.incore.service.space;
 
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
 import org.apache.log4j.Logger;
-
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
 
 
 /**